==> notebooks/01_introduction.ipynb <==
# An Introduction to Machine Learning
In this series of notebooks, we will give a brief overview of machine learning. The focus is to give you the intuition behind important concepts in the field of machine learning, to lay the groundwork for the deep learning course. You are not expected to understand all of the code in the notebooks, but you should have a general idea of what is happening.

## Tools: `sklearn` and `pandas`
For classical machine learning, scikit-learn (or short sklearn) is probably the package that is used most. Here is the documentation: https://scikit-learn.org/stable/. It implements a lot of different methods and models that are very frequently used in practice. There are a lot of [practical examples](https://scikit-learn.org/stable/auto_examples/index.html) and [theoretical explanations](https://scikit-learn.org/stable/user_guide.html) available as well.

For handling data, we will use `pandas` ([link to documentation](https://pandas.pydata.org/docs/getting_started/index.html)), a very useful package that allows to handle data in tables.

For visualizing data, we will use `matplotlib` ([link to documentation](https://matplotlib.org/)).

## Installation

Install these packages now by running the following cell in your Jupyter notebook:

!pip install scikit-learn
!pip install pandas
!pip install matplotlib
!pip install ipympl

Now run this cell. If it runs without an error, the installation process was successful. If not, check your internet connection and redo the step above.

import numpy as np
import pandas as pd
from sklearn import metrics
import matplotlib.pyplot as plt
